pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-socketio>=5.5.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "psutil>=7.0.0",
    "requests>=2.32.3",
    "python-magic>=0.4.27",
    "dnspython>=2.7.0",
    "sqlalchemy>=2.0.41",
]
